Add a warning when affinity passed but not used
When gamma=0 and -af/--affinity path exists and reads ok,
add a warning to state that despite existing affinity matrix
the user is running a vanilla beta-VAE due to gamma=0.

This is useful for future development when we fallback on
existing files in data directory (e.g. affinity found, but
user wants to run a beta-VAE).

Fixes #52
Fixes #51