Convert to a CNN classifier
It is based on the exact Encoder from model_a and model_b.

Loss has been re-writen to include only the regularisation done on
the encoder - so Affinity Loss and KLD Loss. Reconstruction has been
removed.

Pass batch has been altered as no reconstruction is produced,
copy of input is used as a dummy x_recon return but it is
not at any point fed into the reconstruction.